fix: 아이템 목록 조회 페이징 검증 및 정렬 안정성 개선

+++ BE/src/main/java/com/cwww/item/mapper/ItemMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface ItemMapper {
 
     List<Item> findAll();
-    List<Item> find(@Param("size") int size, @Param("offset") int offset);
+    List<Item> find(@Param("size") int size, @Param("offset") long offset);
     int countAll();
 
 }
